+++ examples/openapi-ts-next/src/client/sdk.gen.ts
@@ -52,6 +52,11 @@ export type Options<
    * custom client.
    */
   client?: Client;
+  /**
+   * You can pass arbitrary values through the `meta` object. This can be
+   * used to access values that aren't defined as part of the SDK function.
+   */
+  meta?: Record<string, unknown>;
 };
 
 /**