+++ packages/client-nuxt/src/types.ts
@@ -30,6 +30,11 @@ type WithRefs<TData> = {
     : NonNullable<TData[K]> | Ref<NonNullable<TData[K]>>;
 };
 
+// copied from Nuxt
+export type KeysOf<T> = Array<
+  T extends T ? (keyof T extends string ? keyof T : never) : never
+>;
+
 export interface Config<T extends ClientOptions = ClientOptions>
   extends Omit<
       FetchOptions<unknown>,
@@ -53,6 +58,8 @@ export interface Config<T extends ClientOptions = ClientOptions>
 
 export interface RequestOptions<
   TComposable extends Composable = Composable,
+  ResT = unknown,
+  DefaultT = undefined,
   Url extends string = string,
 > extends Config,
     WithRefs<{
@@ -61,11 +68,11 @@ export interface RequestOptions<
        *
        * {@link https://developer.mozilla.org/docs/Web/API/fetch#body}
        */
-      body?: BodyInit | Record<string, any> | null;
+      body?: unknown;
       path?: FetchOptions<unknown>['query'];
       query?: FetchOptions<unknown>['query'];
     }> {
-  asyncDataOptions?: AsyncDataOptions<unknown>;
+  asyncDataOptions?: AsyncDataOptions<ResT, ResT, KeysOf<ResT>, DefaultT>;
   composable: TComposable;
   key?: string;
   /**
@@ -77,18 +84,18 @@ export interface RequestOptions<
 
 export type RequestResult<
   TComposable extends Composable,
-  TData,
+  ResT,
   TError,
 > = TComposable extends '$fetch'
-  ? ReturnType<typeof $fetch<TData>>
+  ? ReturnType<typeof $fetch<ResT>>
   : TComposable extends 'useAsyncData'
-    ? ReturnType<typeof useAsyncData<TData | null, TError>>
+    ? ReturnType<typeof useAsyncData<ResT | null, TError>>
     : TComposable extends 'useFetch'
-      ? ReturnType<typeof useFetch<TData | null, TError>>
+      ? ReturnType<typeof useFetch<ResT | null, TError>>
       : TComposable extends 'useLazyAsyncData'
-        ? ReturnType<typeof useLazyAsyncData<TData | null, TError>>
+        ? ReturnType<typeof useLazyAsyncData<ResT | null, TError>>
         : TComposable extends 'useLazyFetch'
-          ? ReturnType<typeof useLazyFetch<TData | null, TError>>
+          ? ReturnType<typeof useLazyFetch<ResT | null, TError>>
           : never;
 
 export interface ClientOptions {
@@ -97,20 +104,22 @@ export interface ClientOptions {
 
 type MethodFn = <
   TComposable extends Composable,
-  TData = unknown,
+  ResT = unknown,
   TError = unknown,
+  DefaultT = undefined,
 >(
-  options: Omit<RequestOptions<TComposable>, 'method'>,
-) => RequestResult<TComposable, TData, TError>;
+  options: Omit<RequestOptions<TComposable, ResT, DefaultT>, 'method'>,
+) => RequestResult<TComposable, ResT, TError>;
 
 type RequestFn = <
   TComposable extends Composable,
-  TData = unknown,
+  ResT = unknown,
   TError = unknown,
+  DefaultT = undefined,
 >(
-  options: Omit<RequestOptions<TComposable>, 'method'> &
-    Pick<Required<RequestOptions<TComposable>>, 'method'>,
-) => RequestResult<TComposable, TData, TError>;
+  options: Omit<RequestOptions<TComposable, ResT, DefaultT>, 'method'> &
+    Pick<Required<RequestOptions<TComposable, ResT, DefaultT>>, 'method'>,
+) => RequestResult<TComposable, ResT, TError>;
 
 /**
  * The `createClientConfig()` function will be called on client initialization
@@ -149,7 +158,12 @@ type OmitKeys<T, K> = Pick<T, Exclude<keyof T, K>>;
 export type Options<
   TComposable extends Composable,
   TData extends TDataShape = TDataShape,
-> = OmitKeys<RequestOptions<TComposable>, 'body' | 'path' | 'query' | 'url'> &
+  ResT = unknown,
+  DefaultT = undefined,
+> = OmitKeys<
+  RequestOptions<TComposable, ResT, DefaultT>,
+  'body' | 'path' | 'query' | 'url'
+> &
   WithRefs<Omit<TData, 'url'>>;
 
 export type OptionsLegacyParser<TData = unknown> = TData extends { body?: any }