+++ packages/openapi-ts/src/initConfigs.ts
@@ -28,15 +28,19 @@ const getInput = (userConfig: UserConfig): Config['input'] => {
   };
   if (typeof userConfig.input === 'string') {
     input.path = userConfig.input;
-  } else if (userConfig.input && userConfig.input.path) {
+  } else if (
+    userConfig.input &&
+    (userConfig.input.path || userConfig.input.organization)
+  ) {
     input = {
       ...input,
+      path: 'https://get.heyapi.dev',
       ...userConfig.input,
     };
   } else {
     input = {
       ...input,
-      path: userConfig.input,
+      path: userConfig.input as Record<string, unknown>,
     };
   }
   return input;
@@ -236,10 +240,10 @@ const getWatch = (
 };
 
 export const initConfigs = async (
-  userConfig: UserConfig,
+  userConfig: UserConfig | undefined,
 ): Promise<Config[]> => {
   let configurationFile: string | undefined = undefined;
-  if (userConfig.configFile) {
+  if (userConfig?.configFile) {
     const parts = userConfig.configFile.split('.');
     configurationFile = parts.slice(0, parts.length - 1).join('.');
   }