+++ packages/openapi-ts/src/plugins/@hey-api/client-core/types.d.ts
@@ -43,6 +43,13 @@ export namespace Client {
      */
     bundle?: boolean;
     /**
+     * Should the exports from the generated files be re-exported in the index
+     * barrel file?
+     *
+     * @default false
+     */
+    exportFromIndex?: boolean;
+    /**
      * Name of the generated file.
      *
      * @default 'client'