+++ packages/openapi-ts/src/plugins/@tanstack/vue-query/types.d.ts
@@ -1,6 +1,9 @@
 import type { Plugin } from '../../types';
+import type { TanStackQuery } from '../query-core/types';
 
-export interface Config extends Plugin.Name<'@tanstack/vue-query'> {
+export interface Config
+  extends Plugin.Name<'@tanstack/vue-query'>,
+    TanStackQuery.Config {
   /**
    * Generate {@link https://tanstack.com/query/v5/docs/framework/vue/reference/infiniteQueryOptions `infiniteQueryOptions()`} helpers? These will be generated from GET and POST requests where a pagination parameter is detected.
    *