+++ packages/openapi-ts/src/plugins/zod/config.ts
@@ -6,6 +6,7 @@ export const defaultConfig: Plugin.Config<Config> = {
   _handler: handler,
   _handlerLegacy: () => {},
   _tags: ['validator'],
+  exportFromIndex: false,
   name: 'zod',
   output: 'zod',
 };