+++ packages/openapi-ts/test/__snapshots__/3.1.x/clients/@hey-api/client-fetch/bundle/client/index.d.ts
@@ -162,7 +162,7 @@ interface RequestOptions<ThrowOnError extends boolean = boolean, Url extends str
      *
      * {@link https://developer.mozilla.org/docs/Web/API/fetch#body}
      */
-    body?: RequestInit['body'] | Record<string, unknown> | Array<Record<string, unknown>> | Array<unknown> | number;
+    body?: unknown;
     path?: Record<string, unknown>;
     query?: Record<string, unknown>;
     /**