+++ examples/openapi-ts-tanstack-angular-query-experimental/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @example/openapi-ts-tanstack-angular-query-experimental
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [[`c0b36b9`](https://github.com/hey-api/openapi-ts/commit/c0b36b95645d484034c3af145c5554867568979b)]:
+  - @hey-api/client-fetch@0.8.3
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [[`b8cc9f8`](https://github.com/hey-api/openapi-ts/commit/b8cc9f8a5eaf4f4ff345abc49c14c6b96744c2ea)]:
+  - @hey-api/client-fetch@0.8.2
+
 ## 0.0.17
 
 ### Patch Changes